adds documentation
