+++ core/cu29_export/Cargo.toml
@@ -25,7 +25,6 @@ bincode = { workspace = true }
 pyo3 = { version = "0.25.0", optional = true, features = ["extension-module"] }
 
 [dev-dependencies]
-cu29-log-runtime = { workspace = true }
 tempfile = { workspace = true }
 fs_extra = "1.3.0"
