+++ core/cu29_intern_strs/Cargo.toml
@@ -12,6 +12,10 @@ homepage.workspace = true
 repository.workspace = true
 
 [dependencies]
-cu29-traits = { workspace = true }
+anyhow = { version = "*" }
 rkv = { version = "0.19.0", features = ["lmdb"] }
 byteorder = "1.5.0"
+
+[features]
+# enables a more verbose build log showing the index generation.
+macro_debug = []
